squibler_ai/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
distro==1.9.0
Django==5.1.3
django-channels==0.7.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
idna==3.10
jiter==0.8.0
joblib==1.4.2
kombu==5.4.2
nltk==3.9.1
oauthlib==3.2.2
openai==1.55.3
prompt_toolkit==3.0.48
pydantic==2.10.2
pydantic_core==2.27.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
redis==5.2.0
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
